requirements.txt
Flask==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
line-bot-sdk==3.5.1